requirements.txt
boto3==1.34.0
botocore==1.34.162
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0
flask-restx==0.5.1
werkzeug==3.0.1
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pandas==2.2.0
numpy==1.26.3
pydantic==2.4.2
python-json-logger==2.0.7
colorlog==6.7.0
psutil==5.9.8
cryptography==41.0.4
pyjwt==2.10.1
jinja2==3.1.2
reportlab==4.0.4
openpyxl==3.1.5
requests==2.31.0
httpx==0.25.0
pytest==7.4.4
pytest-cov==4.1.0
responses==0.24.1
factory-boy==3.3.0
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.0
ipython==8.17.2
gunicorn==21.2.0